Changes marathon to use vault for secrets
Secrets are all pushed to vault.
A wrapper script is called from within the systemd service to
extract the values from vault.  Any files are written to private temp
and environment variables are set with vault values.
There is also a token service and timer needed to keep the read only
token alive